Initial tempmut glow commit.
This changes contamination to now cause temporary mutations, with increasingly
worse effects at every glow tier. It also adds a persistent glow effect to a
subset of Charms spells, which causes them to mutate the character for as long
as the spells are active. The experience required to remove temporary muts has
been reduced at lower XLs; it was obvious in testing that the old tempmut
removal formula did not expect the player to be able to rack up tempmuts at
low levels.

This is still very much a work in progress. In particular, I expect that
ABIL_DISPERSE_CHARMS will not remain for long.